refactor userData to fetch from server action; remove profileCardsDataStore
