+++ src/app/(auth)/dashboard/[...cardId]/layout.tsx
@@ -1,5 +1,7 @@
+import { getUserData } from '@/actions/userDataActions';
 import { isAuthenticated } from '@/app/utils/auth/amplify-utils';
 import Header from '@/components/common/Header';
+import { UserData } from '@/types/User.type';
 import { redirect } from 'next/navigation';
 
 export default async function DashboardLayout({
@@ -13,9 +15,11 @@ export default async function DashboardLayout({
   if (!isSignedIn)
     redirect(`/login?redirect=/dashboard/${params.cardId.join('/')}`);
 
+  const userData = await getUserData();
+
   return (
     <>
-      <Header />
+      <Header userData={userData as UserData} />
       {children}
     </>
   );
