+++ src/app/(auth)/dashboard/[...cardId]/not-found.tsx
@@ -2,21 +2,23 @@ import CommonHero from '@/components/common/CommonHero';
 import { Hero } from '@/types/CommonHero.type';
 import Dashboard404 from '@/components/common/Dashboard404';
 import Header from '@/components/common/Header';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
 
-const NotFound = () => {
+export default async function NotFound() {
   const hero: Hero = {
     heading: '',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
       <div>
-        <Header />
+        <Header userData={userData as UserData} />
         <CommonHero hero={hero} />
       </div>
       <Dashboard404 />
     </>
   );
-};
-
-export default NotFound;
+}
