--- src/app/(auth)/profile/tempCodeRunnerFile.tsx
@@ -1,1 +1,0 @@
-<AlertModal />
