+++ src/app/(auth)/settings/layout.tsx
@@ -1,5 +1,7 @@
+import { getUserData } from '@/actions/userDataActions';
 import { isAuthenticated } from '@/app/utils/auth/amplify-utils';
 import Header from '@/components/common/Header';
+import { UserData } from '@/types/User.type';
 import { redirect } from 'next/navigation';
 
 export default async function SettingsLayout({
@@ -10,9 +12,11 @@ export default async function SettingsLayout({
   const { isSignedIn } = await isAuthenticated();
   if (!isSignedIn) redirect('/login?redirect=/settings');
 
+  const userData = await getUserData();
+
   return (
     <>
-      <Header />
+      <Header userData={userData as UserData} />
       {children}
     </>
   );
