+++ src/app/about-us/page.tsx
@@ -10,6 +10,8 @@ import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import { Metadata } from 'next';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
 
 export const metadata: Metadata = {
   title: SEO_CONFIG.aboutUs.title,
@@ -19,7 +21,7 @@ export const metadata: Metadata = {
   },
 };
 
-const AboutUs = () => {
+export default async function AboutUs() {
   // SEO
   const hero: Hero = {
     heading: 'About Us',
@@ -27,11 +29,13 @@ const AboutUs = () => {
     title: 'Your Title Here',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
       <div className="overflow-hidden">
         <div className=" about-page__hero-bg bg-no-repeat bg-cover">
-          <Header />
+          <Header userData={userData as UserData} />
           <CommonHero hero={hero} />
         </div>
         <main>
@@ -44,6 +48,4 @@ const AboutUs = () => {
       </div>
     </>
   );
-};
-
-export default AboutUs;
+}
