+++ src/app/login/page.tsx
@@ -1,15 +1,19 @@
+import { getUserData } from '@/actions/userDataActions';
 import AuthClient from '@/components/auth/AuthClient';
 import Header from '@/components/common/Header';
+import { UserData } from '@/types/User.type';
 import { Suspense } from 'react';
 
-const LoginPage = ({
+export default async function LoginPage({
   searchParams,
 }: {
   searchParams?: { [key: string]: string | undefined };
-}) => {
+}) {
+  const userData = await getUserData();
+
   return (
     <>
-      <Header />
+      <Header userData={userData as UserData} />
       <div className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
         <div className="mt-10">
           <Suspense>
@@ -19,6 +23,4 @@ const LoginPage = ({
       </div>
     </>
   );
-};
-
-export default LoginPage;
+}
