+++ src/app/logout/page.tsx
@@ -3,20 +3,22 @@ import { Hero } from '@/types/CommonHero.type';
 import LoggedOut from '@/components/auth/LoggedOut';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
 
-const LogoutPage = () => {
+export default async function LogoutPage() {
   const hero: Hero = {
     heading: '',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
-    <Header/>
+      <Header userData={userData as UserData} />
       <CommonHero hero={hero} />
       <LoggedOut />
       <Footer />
     </>
   );
-};
-
-export default LogoutPage;
+}
