+++ src/app/not-found.tsx
@@ -5,6 +5,8 @@ import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import { Metadata } from 'next';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
 
 export const metadata: Metadata = {
   title: SEO_CONFIG.notFound.title,
@@ -14,21 +16,21 @@ export const metadata: Metadata = {
   },
 };
 
-const NotFound = () => {
+export default async function NotFound() {
   const hero: Hero = {
     heading: '',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
       <div>
-        <Header />
+        <Header userData={userData as UserData} />
         <CommonHero hero={hero} />
       </div>
       <Custom404 />
       <Footer />
     </>
   );
-};
-
-export default NotFound;
+}
