+++ src/app/sign-up/page.tsx
@@ -1,10 +1,12 @@
 import CommonHero from '@/components/common/CommonHero';
-import { Hero } from '@/types/CommonHero.type';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import SignUpForm from '@/components/sign-up/SignUpForm';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import { Metadata } from 'next';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
+import { Hero } from '@/types/CommonHero.type';
 
 export const metadata: Metadata = {
   title: SEO_CONFIG.signUp.title,
@@ -14,7 +16,7 @@ export const metadata: Metadata = {
   },
 };
 
-const SignUp = () => {
+export default async function SignUp() {
   // SEO
   const hero: Hero = {
     heading: 'Start Your Digital Soccer Card Collection Journey',
@@ -22,10 +24,12 @@ const SignUp = () => {
     title: 'Your Title Here',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
       <div className="signup-page__hero-bg bg-no-repeat bg-cover overflow-hidden">
-        <Header />
+        <Header userData={userData as UserData} />
         <CommonHero hero={hero} />
       </div>
       <main>
@@ -34,6 +38,4 @@ const SignUp = () => {
       <Footer />
     </>
   );
-};
-
-export default SignUp;
+}
