+++ src/app/terms-of-use/page.tsx
@@ -5,6 +5,8 @@ import Header from '@/components/common/Header';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import TermsOfUse from '@/components/terms/TermsOfUse';
 import { Metadata } from 'next';
+import { getUserData } from '@/actions/userDataActions';
+import { UserData } from '@/types/User.type';
 
 export const metadata: Metadata = {
   title: SEO_CONFIG.termsOfUse.title,
@@ -14,14 +16,16 @@ export const metadata: Metadata = {
   },
 };
 
-const TermsOfUsePage = () => {
+export default async function TermsOfUsePage() {
   const hero: Hero = {
     heading: 'Terms of Use',
   };
 
+  const userData = await getUserData();
+
   return (
     <>
-      <Header />
+      <Header userData={userData as UserData} />
       <CommonHero hero={hero} />
       <main>
         <TermsOfUse />
@@ -29,6 +33,4 @@ const TermsOfUsePage = () => {
       <Footer />
     </>
   );
-};
-
-export default TermsOfUsePage;
+}
