+++ src/components/user-portal/CardThumbnail.tsx
@@ -15,19 +15,12 @@ import { useAtomValue, useAtom } from 'jotai';
 import { invitesDataAtom } from '@/states/invitesDataStore';
 import { inviteRevokeActionAtom } from '@/states/InviteRevokeStore';
 import { useRouter } from 'next/navigation';
-import { IProfileProps } from '@/types/Dashboard.type';
-import { GuestCards, OwnedCards } from '@/types/User.type';
+import { ICards } from '@/types/User.type';
 import { Invites } from '@/types/User.type';
 import { sourceSans3 } from '@/app/utils/helpers';
 
-interface ICardData {
-  result: IProfileProps;
-  ownedCardInfo: OwnedCards;
-  guestCardInfo: GuestCards;
-}
-
 interface ICardThumbnailProps {
-  cardData: ICardData;
+  cardData: ICards;
   isOwned: boolean;
   inSettings: boolean;
 }
@@ -80,12 +73,12 @@ const filterAndKeepBestInvite = (
 
 const getFilteredInvites = (
   invites: Invites[],
-  cardData: ICardData,
+  cardData: ICards,
 ): Invites[] => {
   const typeOwnerInvitationFilter = invites.filter(
     (invitation) => invitation.inviter_email !== invitation.guest_email,
   );
-  const cardId = cardData.ownedCardInfo?.card_id;
+  const cardId = cardData.card_id;
   if (cardId) {
     return filterAndKeepBestInvite(typeOwnerInvitationFilter, cardId);
   }
@@ -121,7 +114,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
   const [revokeSubmittedId, setRevokeSubmittedId] = useState<string | null>(
     null,
   );
-  const { name, team, club, card_url, number, club_logo } = cardData.result;
+  const { name, team, club, card_image_url, number, club_logo } = cardData;
   const handleGoToDashboard = (slug: string | null) => {
     //go to dashboard click
     router.push(`/dashboard/${slug}`);
@@ -150,7 +143,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
       inviteRevokeAction({
         action: 'invite',
         guest_email: invitation.email,
-        card_image_id: cardData.ownedCardInfo.card_id,
+        card_image_id: cardData.card_id,
       })
         .then((response) => {
           if (
@@ -258,12 +251,10 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
             <div className="flex justify-start items-start min-w-[260px]">
               <div
                 className="relative w-24 h-28 justify-end cursor-pointer"
-                onClick={() =>
-                  handleGoToDashboard(cardData?.ownedCardInfo.dashboard_slug)
-                }
+                onClick={() => handleGoToDashboard(cardData.dashboard_slug)}
               >
                 <Image
-                  src={card_url as string}
+                  src={card_image_url as string}
                   alt="Card Thumbnail"
                   layout="fill"
                   objectFit="contain"
@@ -328,7 +319,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
                             onClick={() => {
                               triggerRevokeOrInvite(
                                 invite.invite_id,
-                                cardData?.result.name,
+                                cardData.name,
                               );
                             }}
                           >
@@ -463,16 +454,16 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
           <>
             <div className="flex justify-between flex-col md:flex-row items-center w-full">
               <div className="relative w-[300px] h-[350px] md:w-[160px] md:h-[210px]">
-                {card_url ? (
+                {card_image_url ? (
                   <Image
-                    src={card_url}
+                    src={card_image_url}
                     alt="Card Thumbnail"
                     layout="fill"
                     objectFit="contain"
                   />
                 ) : (
                   <Image
-                    src={card_url as string}
+                    src={card_image_url as string}
                     alt="Default Card Thumbnail"
                     layout="fill"
                     objectFit="contain"
@@ -509,9 +500,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
               <div className="flex md:flex-col my-6 md:mx-4 gap-2 items-center justify-center">
                 <button
                   className="text-darkgray w-[140px] md:w-[160px] h-8 bg-skyblue text-xs md:text-sm rounded-full font-normal hover:opacity-90 transform hover:scale-95 ease-in-out"
-                  onClick={() =>
-                    handleGoToDashboard(cardData?.ownedCardInfo.dashboard_slug)
-                  }
+                  onClick={() => handleGoToDashboard(cardData.dashboard_slug)}
                 >
                   go to dashboard
                 </button>
@@ -614,12 +603,10 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
           <div className="flex justify-start items-start min-w-[250px]">
             <div
               className="relative w-24 h-28 justify-end cursor-pointer"
-              onClick={() =>
-                handleGoToDashboard(cardData?.guestCardInfo.dashboard_slug)
-              }
+              onClick={() => handleGoToDashboard(cardData.dashboard_slug)}
             >
               <Image
-                src={card_url as string} // Cast card_url to string
+                src={card_image_url as string}
                 alt="Card Thumbnail"
                 layout="fill"
                 objectFit="contain"
@@ -651,18 +638,18 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
               <div className="flex justify-between items-center">
                 <div className="text-sm flex flex-col">
                   <p className="md:text-center font-extralight">
-                    {cardData?.guestCardInfo?.inviter_email}
+                    {cardData.inviter_email}
                   </p>
                 </div>
                 <div
                   className="flex items-center justify-end cursor-pointer"
                   onClick={() => {
                     triggerDecline(
-                      cardData?.guestCardInfo.invite_id,
-                      cardData?.guestCardInfo.inviter_email,
-                      cardData?.result.name,
+                      cardData.invite_id as string,
+                      cardData.inviter_email,
+                      cardData.name,
                     );
-                    setDeclinedInviteId(cardData?.guestCardInfo.invite_id);
+                    setDeclinedInviteId(cardData.invite_id as string);
                   }}
                 >
                   <div className="text-sm py-4 mx-2 md:mx-4 text-end">
@@ -674,7 +661,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
                   />
                 </div>
               </div>
-              {declinedInviteId === cardData?.guestCardInfo.invite_id && (
+              {declinedInviteId === cardData.invite_id && (
                 <p className="text-xs text-primary mt-1">
                   We will notify the owner you decline this invitation
                 </p>
@@ -711,16 +698,16 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
               )}
             </div>
             <div className="-mt-0 md:-mt-6 w-full md:w-1/2 h-56 md:h-36 relative">
-              {card_url ? (
+              {card_image_url ? (
                 <Image
-                  src={card_url}
+                  src={card_image_url}
                   alt="Card Thumbnail"
                   layout="fill"
                   objectFit="contain"
                 />
               ) : (
                 <Image
-                  src={card_url as string}
+                  src={card_image_url as string}
                   alt="Default Card Thumbnail"
                   layout="fill"
                   objectFit="contain"
@@ -731,9 +718,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
           <div className="w-full flex justify-end pb-4 pr-4">
             <button
               // onClick={triggerRevokeOrInvite}
-              onClick={() =>
-                handleGoToDashboard(cardData?.guestCardInfo.dashboard_slug)
-              }
+              onClick={() => handleGoToDashboard(cardData.dashboard_slug)}
               className="text-darkgray w-[130px] md:w-[160px] h-8 bg-skyblue text-sm rounded-full font-normal hover:opacity-90 transform hover:scale-95 ease-in-out"
             >
               go to dashboard
@@ -746,7 +731,7 @@ const CardThumbnail: React.FC<ICardThumbnailProps> = ({
 };
 
 const RenderCardThumbnail: React.FC<{
-  cardData: ICardData | null | undefined;
+  cardData: ICards | null | undefined;
   isOwned: boolean;
   inSettings: boolean;
 }> = ({ cardData, isOwned, inSettings }) => {
