+++ src/components/user-portal/FAQ.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { FC, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus, faMinus } from '@fortawesome/free-solid-svg-icons';
