+++ src/components/user-portal/ManageReferrals.tsx
@@ -1,29 +1,24 @@
 'use client';
 
-import { OwnedCards, GuestCards, UserData } from '@/types/User.type';
-import RenderCardThumbnail from '@/components/user-portal/CardThumbnail';
-import { useAtomValue } from 'jotai';
 import {
-  ownedCardsDataAtom,
-  guestCardsDataAtom,
-} from '@/states/profileCardsDataStore';
-import { IProfileProps } from '@/types/Dashboard.type';
+  ICards,
+  UserData,
+  emptyOwnedCard,
+  emptyGuestCard,
+} from '@/types/User.type';
+import RenderCardThumbnail from '@/components/user-portal/CardThumbnail';
 import { sourceSans3 } from '@/app/utils/helpers';
-
-interface ICardData {
-  result: IProfileProps;
-  ownedCardInfo: OwnedCards;
-  guestCardInfo: GuestCards;
-}
 
 //const card_url = '/assets/img/png/anderson-card-img.png';
 
 export default function ManageReferrals({ userData }: { userData: UserData }) {
-  const ownedCardsData = useAtomValue(ownedCardsDataAtom) as ICardData[];
-  const guestCardsData = useAtomValue(guestCardsDataAtom) as ICardData[];
+  const ownedCardsData =
+    userData.owned_cards === null ? [emptyOwnedCard] : userData.owned_cards;
+  const guestCardsData =
+    userData.guest_cards === null ? [emptyGuestCard] : userData.guest_cards;
 
   const acceptedGuestCards = guestCardsData.filter(
-    (card) => card.guestCardInfo.status === 'accepted',
+    (card) => card.status === 'accepted',
   );
 
   return (
@@ -46,8 +41,8 @@ export default function ManageReferrals({ userData }: { userData: UserData }) {
         {ownedCardsData.length > 0 ? (
           <div className="flex flex-wrap w-full gap-4">
             {ownedCardsData.map(
-              (cardData: ICardData, idx: React.Key | null | undefined) => {
-                if (!cardData.ownedCardInfo.card_id) {
+              (cardData: ICards, idx: React.Key | null | undefined) => {
+                if (!cardData.card_id) {
                   return (
                     <div
                       key={idx}
@@ -63,7 +58,6 @@ export default function ManageReferrals({ userData }: { userData: UserData }) {
                 return (
                   <RenderCardThumbnail
                     key={idx}
-                    // userData={cardData}
                     cardData={cardData}
                     isOwned={true}
                     inSettings={true}
@@ -90,7 +84,7 @@ export default function ManageReferrals({ userData }: { userData: UserData }) {
         {acceptedGuestCards.length > 0 ? (
           <div className="flex justify-start py-2 overflow-x-auto hide-scrollbar gap-4">
             {acceptedGuestCards.map(
-              (cardData: ICardData, idx: React.Key | null | undefined) => {
+              (cardData: ICards, idx: React.Key | null | undefined) => {
                 if (!cardData) {
                   return null;
                 }
